chore: address requested changes and comments from PR review
